+++ WriteInstinct/OpenCat.h
@@ -69,8 +69,8 @@
 #define I2C_EEPROM //comment this line out if you don't have an I2C EEPROM in your DIY board. 
 
 //postures and movements trained by RongzhongLi
-//#include "InstinctBittle.h" //activate the correct header file according to your model
-#include "InstinctNybble.h"
+#include "InstinctBittle.h" //activate the correct header file according to your model
+//#include "InstinctNybble.h"
 
 //#define NyBoard_V0_1
 //#define NyBoard_V0_2